chore: move utils into `utils` dir
