+++ sui/utils/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+    ...require('./amount-utils'),
+    ...require('./sign-utils'),
+    ...require('./types-utils'),
+    ...require('./utils'),
+    ...require('./upgrade-utils'),
+    ...require("./cli-utils"),
+};
